[Fix]#10 - PageVC 수정
