+++ Weather_App/Weather_App/Detail/DetailView.swift
@@ -14,8 +14,13 @@ class DetailView: UIView {
     
     // MARK: - Properties
     
+    var VCListNum = 0
+    var weatherData = Weathers()
+    
     // MARK: - UI Components
-    let detailCollectionView = UICollectionView(frame: .zero, collectionViewLayout: UICollectionViewLayout())
+    
+    let detailBackImageView = UIImageView()
+    let detailTableView = UITableView()
     
     // MARK: - Life Cycle
     override init(frame: CGRect) {
@@ -35,33 +40,34 @@ class DetailView: UIView {
     // MARK: - Custom Method
     
     private func register() {
-        detailCollectionView.register(DetailCollectionViewCell.self, forCellWithReuseIdentifier: DetailCollectionViewCell.identifier)
+        detailTableView.register(DetailTableViewCell.self, forCellReuseIdentifier: DetailTableViewCell.identifier)
     }
     
     private func style() {
-        detailCollectionView.do {
-            let layout = UICollectionViewFlowLayout()
-            layout.scrollDirection = .horizontal
-            
-            $0.backgroundColor = UIColor(hex: 0x2A3040)
-            $0.collectionViewLayout = layout
-            $0.isPagingEnabled = true
-            $0.decelerationRate = UIScrollView.DecelerationRate.fast
-            $0.isScrollEnabled = true
-            $0.showsHorizontalScrollIndicator = false
-        }
+        self.backgroundColor = UIColor(hex: 0x2A3040)
         
+        detailBackImageView.do {
+            $0.image = .weatherLarge
+        }
+    
+        detailTableView.do {
+            $0.backgroundColor = .clear
+            $0.isScrollEnabled = true
+            $0.showsVerticalScrollIndicator = true
+        }
     }
     
     private func hierarchy() {
-        self.addSubviews(detailCollectionView)
+        self.addSubviews(detailBackImageView, detailTableView)
     }
     
     private func layout() {
-        detailCollectionView.snp.makeConstraints() {
-            $0.top.leading.trailing.equalToSuperview()
-            $0.bottom.equalToSuperview()
+        detailBackImageView.snp.makeConstraints() {
+            $0.edges.equalToSuperview()
+        }
+        
+        detailTableView.snp.makeConstraints() {
+            $0.edges.equalToSuperview()
         }
     }
 }
-
