+++ Weather_App/Weather_App/List/Assets.xcassets/Particle Sprite Atlas.spriteatlas/Contents.json
@@ -1,0 +1,6 @@
+{
+  "info" : {
+    "author" : "xcode",
+    "version" : 1
+  }
+}
